refs #42 -- starting tests for the dir_list method of the builder
Incidentally, moved the Builder() class to the correct module
